# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API para radioxenter",
    "description": "Documentation automatically generated by the <b>swagger.autogen</b> module."
  },
  "host": "localhost:4000",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/usuarios/{cod_usuario}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "description": "Endpoint para obtener un usuario",
        "parameters": [
          {
            "name": "cod_usuario",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/usuarios": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "description": "Endpoint para obtener usuarios",
        "parameters": [],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Usuarios"
        ],
        "description": "Endpoint para registrar un usuario.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "Usuarios"
        ],
        "description": "Endpoint para editar un usuario.",
        "parameters": [
          {
            "description": "description",
            "in": "body",
            "required": true,
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Usuario": {
      "type": "object",
      "properties": {
        "nombres_usuario": {
          "type": "string",
          "example": ""
        },
        "apellidos_usuario": {
          "type": "string",
          "example": ""
        },
        "telefono_usuario": {
          "type": "string",
          "example": ""
        },
        "direccion_usuario": {
          "type": "string",
          "example": ""
        },
        "documento_usuario": {
          "type": "number",
          "example": 0
        },
        "ocupacion_usuario": {
          "type": "string",
          "example": ""
        },
        "fecha_nacimiento_usuario": {
          "type": "string",
          "example": "yyyy-MM-dd"
        },
        "correo_usuario": {
          "type": "string",
          "example": ""
        },
        "genero_usuario": {
          "type": "string",
          "example": ""
        },
        "celular_usuario": {
          "type": "string",
          "example": ""
        },
        "cod_sexo": {
          "type": "number",
          "example": 0
        },
        "cod_tipo_documento": {
          "type": "number",
          "example": 0
        },
        "cod_ciudad": {
          "type": "number",
          "example": 0
        },
        "cod_tipo_pref_entrega": {
          "type": "number",
          "example": 0
        }
      },
      "xml": {
        "name": "Usuario"
      }
    }
  }
}